refactor: 상세 조회 API 응답에서 라벨을 메시지 단위에서 Thread 단위로 이동 (#92)
