+++ core/src/main/java/com/mailsangja/core/dto/inbox/MessageResponse.java
@@ -2,7 +2,6 @@ package com.mailsangja.core.dto.inbox;
 
 import com.mailsangja.db.entity.mail.Direction;
 import com.mailsangja.db.entity.mail.Message;
-import com.mailsangja.db.dto.MessageLabelView;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.time.LocalDateTime;
@@ -39,25 +38,16 @@ public record MessageResponse(
         @Schema(description = "본문 HTML")
         String bodyHtml,
         @Schema(description = "첨부파일 목록")
-        List<AttachmentResponse> attachments,
-        @Schema(description = "메시지에 적용된 라벨 목록")
-        List<LabelSummary> labels
+        List<AttachmentResponse> attachments
 ) {
-    public record LabelSummary(
-            @Schema(description = "라벨 ID") UUID labelId,
-            @Schema(description = "라벨 이름") String name,
-            @Schema(description = "라벨 색상 코드", example = "#FF5733") String colorCode
-    ) {}
-
-    public static MessageResponse from(Message message, Map<String, String> contactNameByEmail, List<MessageLabelView> labelViews) {
-        return from(message, message.getBodyHtml(), contactNameByEmail, labelViews);
+    public static MessageResponse from(Message message, Map<String, String> contactNameByEmail) {
+        return from(message, message.getBodyHtml(), contactNameByEmail);
     }
 
     public static MessageResponse from(
             Message message,
             String renderedBodyHtml,
-            Map<String, String> contactNameByEmail,
-            List<MessageLabelView> labelViews
+            Map<String, String> contactNameByEmail
     ) {
         List<AttachmentResponse> attachmentResponses = message.getAttachments().stream()
                 .map(AttachmentResponse::from)
@@ -75,10 +65,6 @@ public record MessageResponse(
                 contactNameByEmail
         );
 
-        List<LabelSummary> labels = labelViews.stream()
-                .map(v -> new LabelSummary(v.labelId(), v.labelName(), v.colorCode()))
-                .toList();
-
         return new MessageResponse(
                 message.getId(),
                 message.getGmailMessageId(),
@@ -93,8 +79,7 @@ public record MessageResponse(
                 message.getSentAt(),
                 message.getBodyText(),
                 renderedBodyHtml,
-                attachmentResponses,
-                labels
+                attachmentResponses
         );
     }
