+++ core/src/main/java/com/mailsangja/core/dto/inbox/ThreadDetailResponse.java
@@ -2,7 +2,7 @@ package com.mailsangja.core.dto.inbox;
 
 import com.mailsangja.db.entity.mail.Message;
 import com.mailsangja.db.entity.mail.Thread;
-import com.mailsangja.db.dto.MessageLabelView;
+import com.mailsangja.db.dto.ThreadMessageLabelView;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.time.LocalDateTime;
@@ -24,23 +24,34 @@ public record ThreadDetailResponse(
         boolean isRead,
         @Schema(description = "최신 메시지 시각")
         LocalDateTime lastMessageAt,
+        @Schema(description = "스레드에 적용된 라벨 목록")
+        List<LabelSummary> labels,
         @Schema(description = "스레드 내 메시지 목록 (sentAt ASC 정렬)")
         List<MessageResponse> messages
 ) {
+    public record LabelSummary(
+            @Schema(description = "라벨 ID") UUID labelId,
+            @Schema(description = "라벨 이름") String name,
+            @Schema(description = "라벨 색상 코드", example = "#FF5733") String colorCode
+    ) {}
+
     public static ThreadDetailResponse from(
             Thread thread,
             List<Message> messages,
             Map<String, String> contactNameByEmail,
-            Map<UUID, List<MessageLabelView>> messageLabelsByMessageId,
+            List<ThreadMessageLabelView> labelViews,
             Map<UUID, String> renderedBodyHtmlByMessageId
     ) {
         List<MessageResponse> messageResponses = messages.stream()
                 .map(message -> MessageResponse.from(
                         message,
                         renderedBodyHtmlByMessageId.getOrDefault(message.getId(), message.getBodyHtml()),
-                        contactNameByEmail,
-                        messageLabelsByMessageId.getOrDefault(message.getId(), List.of())
+                        contactNameByEmail
                 ))
+                .toList();
+
+        List<LabelSummary> labels = labelViews.stream()
+                .map(v -> new LabelSummary(v.labelId(), v.labelName(), v.colorCode()))
                 .toList();
 
         return new ThreadDetailResponse(
@@ -50,6 +61,7 @@ public record ThreadDetailResponse(
                 thread.getLatestSubject(),
                 thread.isRead(),
                 thread.getLastMessageAt(),
+                labels,
                 messageResponses
         );
     }
