+++ core/src/main/java/com/mailsangja/core/dto/inbox/ThreadDetailResult.java
@@ -2,15 +2,14 @@ package com.mailsangja.core.dto.inbox;
 
 import com.mailsangja.db.entity.mail.Message;
 import com.mailsangja.db.entity.mail.Thread;
-import com.mailsangja.db.dto.MessageLabelView;
+import com.mailsangja.db.dto.ThreadMessageLabelView;
 
 import java.util.List;
 import java.util.Map;
-import java.util.UUID;
 
 public record ThreadDetailResult(
         Thread thread,
         List<Message> messages,
         Map<String, String> contactNameByEmail,
-        Map<UUID, List<MessageLabelView>> messageLabelsByMessageId
+        List<ThreadMessageLabelView> labels
 ) {}
