+++ core/src/main/java/com/mailsangja/core/dto/trash/TrashThreadDetailResponse.java
@@ -3,7 +3,7 @@ package com.mailsangja.core.dto.trash;
 import com.mailsangja.core.dto.inbox.MessageResponse;
 import com.mailsangja.db.entity.mail.Message;
 import com.mailsangja.db.entity.mail.Thread;
-import com.mailsangja.db.dto.MessageLabelView;
+import com.mailsangja.db.dto.ThreadMessageLabelView;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.time.LocalDateTime;
@@ -25,23 +25,34 @@ public record TrashThreadDetailResponse(
         boolean isRead,
         @Schema(description = "최신 메시지 시각")
         LocalDateTime lastMessageAt,
+        @Schema(description = "스레드에 적용된 라벨 목록")
+        List<LabelSummary> labels,
         @Schema(description = "삭제된 메시지 목록 (sentAt ASC 정렬)")
         List<MessageResponse> messages
 ) {
+    public record LabelSummary(
+            @Schema(description = "라벨 ID") UUID labelId,
+            @Schema(description = "라벨 이름") String name,
+            @Schema(description = "라벨 색상 코드", example = "#FF5733") String colorCode
+    ) {}
+
     public static TrashThreadDetailResponse from(
             Thread thread,
             List<Message> messages,
             Map<String, String> contactNameByEmail,
-            Map<UUID, List<MessageLabelView>> messageLabelsByMessageId,
+            List<ThreadMessageLabelView> labelViews,
             Map<UUID, String> renderedBodyHtmlByMessageId
     ) {
         List<MessageResponse> messageResponses = messages.stream()
                 .map(message -> MessageResponse.from(
                         message,
                         renderedBodyHtmlByMessageId.getOrDefault(message.getId(), message.getBodyHtml()),
-                        contactNameByEmail,
-                        messageLabelsByMessageId.getOrDefault(message.getId(), List.of())
+                        contactNameByEmail
                 ))
+                .toList();
+
+        List<LabelSummary> labels = labelViews.stream()
+                .map(v -> new LabelSummary(v.labelId(), v.labelName(), v.colorCode()))
                 .toList();
 
         return new TrashThreadDetailResponse(
@@ -51,6 +62,7 @@ public record TrashThreadDetailResponse(
                 thread.getLatestSubject(),
                 thread.isRead(),
                 thread.getLastMessageAt(),
+                labels,
                 messageResponses
         );
     }
